Extract existsMessage() function in separate module
